fix(tooltip): get rid of blinking from interactive tooltips (#1617)
Reduce the next delay when tooltip is interactive to prevent poor UX when tooltip content is
hovered. 

When cursor is moved from the tooltip trigger to the content, the tooltip starts to hide 
only to reappear back as soon as the cursor lands on the content. Having the next delay in this
scenario causes blinking as we need to wait those 200ms or 500ms again.