fix: remove material-ui/styles from dependencies (#2165)
Remove dependency on  "@material-ui/styles": "4.10.0" because we have this package installed already with "@material-ui/core" and not pinned to 4.10. This causes possible clash of dependencies, the project may depend on 2 version:
 - "@material-ui/styles": "4.10.0" from Picasso
 - "@material-ui/styles": "4.11.4" from material-ui/core
 
 Styles created with 4.10 can't access the theme created with 4.11.4 and crash.

This dependency is not needed as the material-ui core exports the same hooks, which are always compatible with it, so instead of that additional dependency we should have fixed the imports instead.