refactor: prototype
Currently toolbar is customized and satisfies the design in the
Abstract, but there are two missing parts, one is state of buttons are
not yet implemented (hover and active). And the second is to generate
icons like in the design.

Also, as we are using custom toolbar, toolbar actions are broken, need
to re-implement logic.